+++ src/features/home/components/VersionStatus.tsx
@@ -1,10 +1,27 @@
 import { Check, Download } from "lucide-solid";
 import { createSignal, onMount, Show } from "solid-js";
-import { getLatestVersion } from "@/utils/versionCheck";
+import { getCurrentVersion, getLatestVersion } from "@/utils/versionCheck";
 
 export const VersionStatus = () => {
   const [latestVersion, setLatestVersion] = createSignal<string | null>(null);
   const [versionChecked, setVersionChecked] = createSignal(false);
+
+  const isUpdateAvailable = () => {
+    const latest = latestVersion();
+    return latest !== null && latest !== getCurrentVersion();
+  };
+
+  const handleInstallUpdate = () => {
+    const message = [
+      `A new version (${latestVersion()}) is available.`,
+      "The app will close, update, and reopen automatically.",
+      "Do you want to proceed?",
+    ].join("\n\n");
+    const confirmed = confirm(message);
+    if (confirmed) {
+      window.electronAPI.installUpdate();
+    }
+  };
 
   onMount(() => {
     const checkUpdates = async () => {
@@ -19,7 +36,7 @@ export const VersionStatus = () => {
     <Show when={versionChecked()}>
       <div class="alert alert-soft max-w-150 mx-auto">
         <Show
-          when={latestVersion()}
+          when={isUpdateAvailable()}
           fallback={
             <>
               <Check size={32} />
@@ -48,14 +65,9 @@ export const VersionStatus = () => {
             <button
               type="button"
               class="btn btn-sm btn-primary"
-              onClick={() =>
-                window.open(
-                  "https://github.com/Jordan-Kowal/click-launch/releases/latest",
-                  "_blank",
-                )
-              }
+              onClick={handleInstallUpdate}
             >
-              Download
+              Upgrade
             </button>
             <button
               type="button"